+++ addons/docs/src/frameworks/react/extractProps.ts
@@ -1,9 +1,10 @@
 import PropTypes from 'prop-types';
 import { isForwardRef, isMemo } from 'react-is';
 import { PropDef } from '@storybook/components';
-import { hasDocgen, extractPropsFromDocgen, PropsExtractor, TypeSystem } from '../../lib/docgen';
+import { hasDocgen, extractComponentProps, PropsExtractor, TypeSystem } from '../../lib/docgen';
 import { Component } from '../../blocks/shared';
 import { enhancePropTypesProps } from './propTypes/handleProp';
+import { enhanceTypeScriptProps } from './typeScript/handleProp';
 
 export interface PropDefMap {
   [p: string]: PropDef;
@@ -32,16 +33,19 @@ function getPropDefs(component: Component, section: string): PropDef[] {
     }
   }
 
-  const extractedProps = extractPropsFromDocgen(processedComponent, section);
+  const extractedProps = extractComponentProps(processedComponent, section);
   if (extractedProps.length === 0) {
     return [];
   }
 
-  if (extractedProps[0].typeSystem === TypeSystem.JAVASCRIPT) {
-    return enhancePropTypesProps(extractedProps, component);
+  switch (extractedProps[0].typeSystem) {
+    case TypeSystem.JAVASCRIPT:
+      return enhancePropTypesProps(extractedProps, component);
+    case TypeSystem.TYPESCRIPT:
+      return enhanceTypeScriptProps(extractedProps);
+    default:
+      return extractedProps.map(x => x.propDef);
   }
-
-  return extractedProps.map(x => x.propDef);
 }
 
 export const extractProps: PropsExtractor = component => ({