+++ addons/docs/src/lib/docgen/createPropDef.ts
@@ -1,16 +1,29 @@
 import { isNil } from 'lodash';
-import { PropDef } from '@storybook/components';
-import { TypeSystem, DocgenInfo, DocgenType } from './types';
+import { PropDef, PropDefaultValue } from '@storybook/components';
+import { TypeSystem, DocgenInfo, DocgenType, DocgenPropDefaultValue } from './types';
 import { JsDocParsingResult } from '../jsdocParser';
-import { createDefaultValue } from './createDefaultValue';
 import { createSummaryValue } from '../utils';
 import { createFlowPropDef } from './flow/createPropDef';
+import { isDefaultValueBlacklisted } from './utils/defaultValue';
+import { createTsPropDef } from './typeScript/createPropDef';
 
 export type PropDefFactory = (
   propName: string,
   docgenInfo: DocgenInfo,
   jsDocParsingResult?: JsDocParsingResult
 ) => PropDef;
+
+function createDefaultValue(defaultValue: DocgenPropDefaultValue): PropDefaultValue {
+  if (!isNil(defaultValue)) {
+    const { value } = defaultValue;
+
+    if (!isDefaultValueBlacklisted(value)) {
+      return createSummaryValue(value);
+    }
+  }
+
+  return null;
+}
 
 function createBasicPropDef(name: string, type: DocgenType, docgenInfo: DocgenInfo): PropDef {
   const { description, required, defaultValue } = docgenInfo;
@@ -57,7 +70,7 @@ export const javaScriptFactory: PropDefFactory = (propName, docgenInfo, jsDocPar
 };
 
 export const tsFactory: PropDefFactory = (propName, docgenInfo, jsDocParsingResult) => {
-  const propDef = createBasicPropDef(propName, docgenInfo.tsType, docgenInfo);
+  const propDef = createTsPropDef(propName, docgenInfo);
 
   return applyJsDocResult(propDef, jsDocParsingResult);
 };