+++ addons/docs/src/lib/docgen/flow/createType.ts
@@ -1,4 +1,4 @@
-import { PropType, PropSummaryValue } from '@storybook/components';
+import { PropType } from '@storybook/components';
 import { isNil } from 'lodash';
 import { DocgenFlowType } from '../types';
 import { createSummaryValue, isTooLongForTypeSummary } from '../../utils';
@@ -12,7 +12,7 @@ interface DocgenFlowUnionType extends DocgenFlowType {
   elements: { name: string; value: string }[];
 }
 
-function generateUnion({ name, raw, elements }: DocgenFlowUnionType): PropSummaryValue {
+function generateUnion({ name, raw, elements }: DocgenFlowUnionType): PropType {
   if (!isNil(raw)) {
     return createSummaryValue(raw);
   }
@@ -24,7 +24,7 @@ function generateUnion({ name, raw, elements }: DocgenFlowUnionType): PropSummar
   return createSummaryValue(name);
 }
 
-function generateFuncSignature({ type, raw }: DocgenFlowType) {
+function generateFuncSignature({ type, raw }: DocgenFlowType): PropType {
   if (!isNil(raw)) {
     return createSummaryValue(raw);
   }
@@ -32,7 +32,7 @@ function generateFuncSignature({ type, raw }: DocgenFlowType) {
   return createSummaryValue(type);
 }
 
-function generateObjectSignature({ type, raw }: DocgenFlowType) {
+function generateObjectSignature({ type, raw }: DocgenFlowType): PropType {
   if (!isNil(raw)) {
     return !isTooLongForTypeSummary(raw) ? createSummaryValue(raw) : createSummaryValue(type, raw);
   }
@@ -40,13 +40,13 @@ function generateObjectSignature({ type, raw }: DocgenFlowType) {
   return createSummaryValue(type);
 }
 
-function generateSignature(flowType: DocgenFlowType): PropSummaryValue {
+function generateSignature(flowType: DocgenFlowType): PropType {
   const { type } = flowType;
 
   return type === 'object' ? generateObjectSignature(flowType) : generateFuncSignature(flowType);
 }
 
-function generateDefault({ name, raw }: DocgenFlowType): PropSummaryValue {
+function generateDefault({ name, raw }: DocgenFlowType): PropType {
   if (!isNil(raw)) {
     return !isTooLongForTypeSummary(raw) ? createSummaryValue(raw) : createSummaryValue(name, raw);
   }