+++ lib/components/src/blocks/DocsPage.tsx
@@ -2,7 +2,7 @@ import React, { FunctionComponent } from 'react';
 import { styled, Theme } from '@storybook/theming';
 import { transparentize } from 'polished';
 
-import { withReset } from '../typography/withReset';
+import { withReset } from '../typography/shared';
 
 const breakpoint = 600;