+++ lib/components/src/blocks/PropsTable/PropJsDoc.tsx
@@ -2,21 +2,21 @@ import React, { FC } from 'react';
 import { styled } from '@storybook/theming';
 import { isNil } from 'lodash';
 import { JsDocTags } from './PropDef';
+import { codeCommon } from '../../typography/shared';
 
 interface PropJsDocProps {
   tags: JsDocTags;
 }
 
-export const Table = styled.table<{}>(({ theme }) => ({
+export const Table = styled.table(({ theme }) => ({
   '&&': {
-    // Resets for cascading/system styles
+    // Escape default table styles
     borderCollapse: 'collapse',
     borderSpacing: 0,
-    color: theme.color.defaultText,
     border: 'none',
 
     tr: {
-      border: 'none',
+      border: 'none !important',
       background: 'none',
     },
 
@@ -25,12 +25,7 @@ export const Table = styled.table<{}>(({ theme }) => ({
       border: 'none',
       width: 'auto!important',
     },
-    // End Resets
-
-    fontSize: theme.typography.size.s2,
-    lineHeight: '20px',
-    textAlign: 'left',
-    width: '100%',
+    // End escape
 
     marginTop: '0',
     marginBottom: '0',
@@ -40,69 +35,30 @@ export const Table = styled.table<{}>(({ theme }) => ({
     },
 
     'th:last-of-type, td:last-of-type': {
-      paddingRight: '20px',
-      width: '20%',
-    },
-
-    th: {
-      paddingTop: 10,
-      paddingBottom: 10,
-
-      '&:not(:first-of-type)': {
-        paddingLeft: 12,
-        paddingRight: 12,
-      },
+      paddingRight: 0,
     },
 
     td: {
-      paddingTop: '8px',
-      paddingBottom: '8px',
+      paddingTop: '0',
+      paddingBottom: '4px',
 
       '&:not(:first-of-type)': {
-        paddingLeft: 12,
-        paddingRight: 12,
+        paddingLeft: 10,
+        paddingRight: 0,
       },
-
-      '&:last-of-type': {
-        paddingRight: '20px',
-      },
-    },
-
-    'tr:first-child': {
-      'td:first-child, th:first-child': {
-        borderTopLeftRadius: 0,
-      },
-      'td:last-child, th:last-child': {
-        borderTopRightRadius: 0,
-      },
-    },
-
-    code: {
-      backgroundColor: theme.color.lighter,
-      padding: '4px',
-      borderRadius: '4px',
-      color: theme.color.darker,
-      fontSize: theme.typography.size.s2,
     },
 
     tbody: {
       boxShadow: 'none',
       border: 'none',
+    },
 
-      tr: {
-        background: 'transparent',
-        overflow: 'hidden',
-        '&:not(:first-child)': {
-          borderTopWidth: 0,
-        },
-      },
+    code: codeCommon({ theme }),
 
-      td: {
-        background: theme.background.content,
-        border: 'none',
-        width: 'auto',
-        color: theme.color.darker,
-      },
+    '& code': {
+      lineHeight: '18px',
+      margin: 0,
+      display: 'inline-block',
     },
   },
 }));